Refactor ReachCodec to be more stream oriented
Now, instead of the responsibility of passing the audio data
falling on the React application, that responsibility lies in
ReachCodec. This is mostly due to limitations with the
emscripten_fiber API, but probably results in a better design,
anyway. The UI initializes a series of objects that connect together
to stream data through a series of transformations before finally
reaching the Audio Device. This design is likened to the Node.js
stream.Transform API.